+++ include/fem-dem/cfd_dem_coupling.h
@@ -96,6 +96,17 @@ private:
   read_dem();
 
   /**
+   * @brief Returns a vector of references to TableHandler objects that need to
+   * be serialized/deserialized for a given solver.
+   *
+   * @return Structure containing a vector of references to TableHandler objects
+   * that needs to be serialized/deserialized for a given solver, and their
+   * corresponding file names.
+   */
+  std::vector<OutputStructTableHandler>
+  gather_tables() override;
+
+  /**
    * @brief Write the CFD-DEM restart files.
    */
   void
@@ -223,7 +234,8 @@ private:
   double                                     smallest_contact_search_criterion;
 
   /**
-   * @brief Class object to store the particle id, boundary id, velocity tensor, omega tensor and time of ongoing collisions.
+   * @brief Class object to store the particle id, boundary id, velocity tensor,
+   *  omega tensor and time of ongoing collisions.
    */
   OngoingCollisionLog<dim> ongoing_collision_log;