+++ include/solvers/output_struct.h
@@ -113,4 +113,30 @@ using OutputStruct = std::variant<OutputStructPostprocessor<dim, VectorType>,
                                   OutputStructSolution<dim, VectorType>,
                                   OutputStructCellVector>;
 
+/**
+ * @brief Struct containing information about the table handlers to be
+ * written/read. It is used to pass a vector of references to TableHandler
+ * objects that should be serialized/deserialized at the writing/reading of a
+ * checkpoint by the solver used, and a vector of references to strings
+ * containing the names of the corresponding files.
+ */
+struct OutputStructTableHandler
+{
+  /**
+   * @brief Constructor
+   *
+   * @param[in] table A reference to a TableHandler object that needs to be
+   * serialized/ deserialized.
+   * @param[in] table_filename A string with the corresponding file name.
+   */
+  OutputStructTableHandler(TableHandler      &table,
+                           const std::string &table_filename)
+    : table(table)
+    , table_filename(table_filename)
+  {}
+  /// Reference to the TableHandler object to be serialized/deserialized
+  TableHandler &table;
+  /// Name of the file associated with the TableHandler for input/output
+  const std::string table_filename;
+};
 #endif